[client/rest]: move Symbol rosetta implementation into symbol subfolder
 problem: Symbol and NEM rosetta implementations should be separated
solution: move Symbol implementation to symbol subfolder
