+++ client/rest/resources/rest.json
@@ -32,7 +32,6 @@
     "transfer"
   ],
   "routeExtensions": [
-    "rosetta"
   ],
   "db": {
     "url": "mongodb://localhost:27017/",
