+++ client/rest/resources/rest.rosetta.json
@@ -15,7 +15,7 @@
 
   "restEndpoint": "http://localhost:3000",
   "routeExtensions": [
-    "rosetta"
+    "symbolRosetta"
   ],
 
   "throttling": {
