+++ client/rest/src/plugins/rosetta/symbol/CatapultProxy.js
