+++ client/rest/src/plugins/routeSystem.js
@@ -29,7 +29,7 @@ import multisig from './multisig/multisig.js';
 import namespace from './namespace/namespace.js';
 import receipts from './receipts/receipts.js';
 import restrictions from './restrictions/restrictions.js';
-import rosetta from './rosetta/rosetta.js';
+import symbolRosetta from './rosetta/symbol/rosetta.js';
 import MessageChannelBuilder from '../connection/MessageChannelBuilder.js';
 import { NetworkLocator } from 'symbol-sdk';
 import { Network } from 'symbol-sdk/symbol';
@@ -48,8 +48,8 @@ const plugins = {
 	restrictions,
 	transfer: empty,
 
-	// other
-	rosetta
+	// rosetta
+	symbolRosetta
 };
 
 export default {
