+++ client/rest/test/plugins/rosetta/symbol/mempoolRoutes_spec.js
@@ -26,13 +26,13 @@ import {
 	assertRosettaErrorRaisedBasicWithRoutes,
 	assertRosettaSuccessBasicWithRoutes
 } from './utils/rosettaTestUtils.js';
-import { convertTransactionSdkJsonToRestJson } from '../../../src/plugins/rosetta/OperationParser.js';
-import mempoolRoutes from '../../../src/plugins/rosetta/mempoolRoutes.js';
-import MempoolResponse from '../../../src/plugins/rosetta/openApi/model/MempoolResponse.js';
-import MempoolTransactionResponse from '../../../src/plugins/rosetta/openApi/model/MempoolTransactionResponse.js';
-import Transaction from '../../../src/plugins/rosetta/openApi/model/Transaction.js';
-import TransactionIdentifier from '../../../src/plugins/rosetta/openApi/model/TransactionIdentifier.js';
-import { RosettaErrorFactory } from '../../../src/plugins/rosetta/rosettaUtils.js';
+import MempoolResponse from '../../../../src/plugins/rosetta/openApi/model/MempoolResponse.js';
+import MempoolTransactionResponse from '../../../../src/plugins/rosetta/openApi/model/MempoolTransactionResponse.js';
+import Transaction from '../../../../src/plugins/rosetta/openApi/model/Transaction.js';
+import TransactionIdentifier from '../../../../src/plugins/rosetta/openApi/model/TransactionIdentifier.js';
+import { convertTransactionSdkJsonToRestJson } from '../../../../src/plugins/rosetta/symbol/OperationParser.js';
+import mempoolRoutes from '../../../../src/plugins/rosetta/symbol/mempoolRoutes.js';
+import { RosettaErrorFactory } from '../../../../src/plugins/rosetta/symbol/rosettaUtils.js';
 import { SymbolFacade, generateMosaicAliasId } from 'symbol-sdk/symbol';
 
 describe('mempool routes', () => {
