+++ client/rest/test/plugins/routeSystem_spec.js
@@ -52,8 +52,8 @@ describe('route system', () => {
 			'restrictions',
 			'transfer',
 
-			// - other
-			'rosetta'
+			// - rosetta
+			'symbolRosetta'
 		]);
 	});
